Move volume api types into the volume component.
Also move volume/client to volume/command
