+++ cli/command/formatter/volume.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/components/volume/types"
 )
 
 const (
