+++ cli/command/formatter/volume_test.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/components/volume/types"
 	"github.com/docker/docker/pkg/stringid"
 	"github.com/docker/docker/pkg/testutil/assert"
 )
