+++ client/volume_inspect.go
@@ -6,7 +6,7 @@ import (
 	"io/ioutil"
 	"net/http"
 
-	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/components/volume/types"
 	"golang.org/x/net/context"
 )
