+++ client/volume_list.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 	"net/url"
 
-	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/filters"
+	"github.com/docker/docker/components/volume/types"
 	"golang.org/x/net/context"
 )
