+++ components/volume/command/list.go
@@ -1,14 +1,14 @@
-package client
+package command
 
 import (
 	"sort"
 
 	"golang.org/x/net/context"
 
-	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/cli"
 	"github.com/docker/docker/cli/command"
 	"github.com/docker/docker/cli/command/formatter"
+	"github.com/docker/docker/components/volume/types"
 	"github.com/docker/docker/opts"
 	"github.com/spf13/cobra"
 )
